refactor(*): group argument relates component
